+++ lib/kubernetes-deploy/deploy_task.rb
@@ -89,13 +89,15 @@ module KubernetesDeploy
 
     NOT_FOUND_ERROR = 'NotFound'
 
-    def initialize(namespace:, context:, current_sha:, template_dir:, logger:, kubectl_instance: nil, bindings: {})
+    def initialize(namespace:, context:, current_sha:, template_dir:, logger:, kubectl_instance: nil, bindings: {},
+      max_watch_seconds: nil)
       @namespace = namespace
       @context = context
       @current_sha = current_sha
       @template_dir = File.expand_path(template_dir)
       @logger = logger
       @kubectl = kubectl_instance
+      @max_watch_seconds = max_watch_seconds
       @renderer = KubernetesDeploy::Renderer.new(
         current_sha: @current_sha,
         template_dir: @template_dir,
@@ -377,7 +379,7 @@ module KubernetesDeploy
 
       if verify
         watcher = ResourceWatcher.new(resources: resources, sync_mediator: @sync_mediator,
-          logger: @logger, deploy_started_at: deploy_started_at)
+          logger: @logger, deploy_started_at: deploy_started_at, timeout: @max_watch_seconds)
         watcher.run(record_summary: record_summary)
       end
     end