+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ### Unreleased
+*Features*
+- Added `--max-watch-seconds=seconds` to kubernetes-restart and kubernetes-deploy. When set
+a timeout error is raised if it takes longer than _seconds_ for any resource to deploy.
+
 *Enhancements*
 - Prune resource quotas ([#264](https://github.com/Shopify/kubernetes-deploy/pull/264/files))